backend/requirements.txt
langgraph
langchain
fastapi
uvicorn[standard]
python-dotenv
langchain_openai
pytest
httpx
requests
